requirements.txt
# Основные зависимости для Telegram GPT Bot
python-telegram-bot==21.11.1
openai==1.64.0
redis==5.2.1
python-dotenv==1.0.1
aiohttp==3.10.11
aiosqlite==0.20.0
